Merged PR 1116488: Create a real native nuget package
Verified that it builds and releases successfully and downloaded the package and installed it on a dummy project in VS.
Added Microsoft.xPlat.AppxPackaging.targets which allows everyone that installs the package to already have our headers in their include path.
Move stuff around in our CMakeLists.txt so we can complete the NUSPEC on clean runs without depending on an specific target.

Structure of the nuget package
- build
-- native
---MACOS
---WIN32
---Microsoft.xPlat.AppxPackaging.targets