Merged PR 1116842: Merge bug_14681942 to feature
Basically, we needed to implement Zip64 support as HelloWorld.appx (SDX) was, under the covers built with a Zip64-based OPC container.  None of our prior collateral was built as such. Also, we didn't have the ability for a client to override unpacking an unsigned appx package (which was blocking end-to-end validation of HelloWorld.appx as it was not signed).

Related work items: #14681942