Merged PR 1161737: Lighten up a few overzealous format restrictions and correctly handle the case
Lighten up a few overzealous format restrictions and correctly handle the case where we do not have a Zip64Locator record and Zip64EndOfCentralDirectory record.  Also, reorder a few of the XML error facilities and conditions to make the code considerably easier to debug on Debug builds.

Related work items: #14879794