Merged PR 1384642: Unpack store signed appx packages on Win7
Manually check for the SHA256 public key of the Microsoft application roots key identifier. This is needed for Win7 as CertVerifyCertificateChainPolicy fails with CERT_E_UNTRUSTEDROOT because it compares against root hashes imbedded in crypt32.dll. In this case, the check is for the Win7 version of crypt32.dll which doesn't have the Microsoft Product root hashes.

Verified on Win7 by running all our test and verifying they passed properly.

Related work items: #15585823