Merged PR 1862819: Support having more than one language on Android for devices running on level 24 or higher.
Android added, on API level 24 support for having more than one language on the device . MSIX min API level is 19, so we can't use Configuration.getLocales() to obtain them and we have to use the deprecated attribute Configuration.locale. We should support returning multiple languages for devices were it is supported and only one when is not possible.

Also add documentation on the README about Android support and removed expecting always $ANDROID_HOME as an environment variable by adding a new switch on CMake that expects the SDK path.

There isn't a problem in building a jar with a higher API level and then run it on a lower API level device. However, there is a problem in building the native lib on a higher API level than one of the device. This is why the distinction has to be made between the NDK and SDK level. In less words, the NDK minimum is 19 and the SDK minimum is 24.

Related work items: #17634105