Merged PR 2229038:  Double dereference crash in function std::pair<bool, Variant> GetAttribute(XmlAttributeName attribute) in xplatappx\src\msix\PAL\XML\msxml6\XmlObject.cpp
Make MSXMLElement GetAttribute simplier and add move constructor to the Variant class. The lack of a move constructor for this class made the copy of the Variant just point to the old VARIANT that got clean in the destructor of the original object. All this time we are just getting lucky (arguably we will always to be lucky).

Related work items: #18688778