Merged PR 2287191: Remove IAppxFile from StreamBase
Remove IAppxFile from StreamBase. Most of the derived classed of StreamBase don't care about if it is an IAppxFile or not and we end up with a lot of them just returning E_NOTIMPL. This is also confusing sometimes a stream can QI for IAppxFile and succeed, but if we ever get an outside stream and asume that is an appx file we will have problems.

IAppxFileInternal had nothing to do with AppxFiles, so I remove it. Its only functionality was to return the size of the stream in the zip container to validate the information of the block map. IStreamInternal::GetSizeOnZip provides this functionality now.

Because the vtables are smaller, this change also reduce the size of the binary ~7k

Related work items: #18752968, #18853644