Merged PR 2300105: switch to use inbox zlib on android.
updating android to use the ndk version of libz.so instead of the static lib built in the xplatappx repo. If you want to use the xplatappx one instead, you can specify the -xzlib build option to makeaosp.sh
Refactored the inflatestream code to keep the common portions of stream state management and pulled out the compression implementation specifics into a compressionObject PAL implementation.
This reduces the binary size by ~30K on android.

Related work items: #18607665