Merged PR 2438384: [MSIX SDK] simplifying the template metaprograming flow in zipobject/objectbase
The current implementation used static polymorphism via templates to scope validation and read to each field, and in some cases defer to the parent container for additional inter-field validations. This made the code more complex than it needs to be.

Simplified that logic by making the structured object to be a sequence of fields that it then reads and validates sequentially.
This reduced the binary size by small amount ~6k.

Related work items: #19248506