Merged PR 2478080: ignoring harmless z_buf_error during zip inflate.
during the refactoring, I noticed that existing code was not including z_buf_error as a error condition. I thought it was unintentional and so added it. Apparently, it is intentionally ignored in msix sdk as well as appxpackaging code.
Added the appx to the  suite of tests for future regression testing.

Related work items: #19340511