Merged PR 2607225: [MSIX SDK] Verify that extracting a package where the files inside it have intl character works
When a package contains files with international characters, the opc container will have them as percentage encoded, but the block map not, this causes the extraction to fail because the files in the opc container and the block map don't match as our decoding/encoding was only for ascii only. This changes expands DecodeFileName and EncodeFileName for non-ascii characters.

Related work items: #19276513