Merged PR 2684301: [MSIX SDK][ENG] Create script to build and run iOS test
This script will run the iOS BVTs automatically.

Its defaults settings test the app in an iPhone 8 emulator for iphonesimulator11.4.
It allows overriding the device and sdk, if we ever need to test it on a specific device/iOS version.
It also allows overriding the project and output file, so we can reuse it if we ever have another iOS test app (API tests?)

Lets never open Xcode again (hooray! :D)

Related work items: #18826750