Merged PR 2807884: Improve API tests and port them to the AOSP and iOS
Re-wrote the entire API tests to not hardcode one package and one bundle and make them compatible with AOSP and iOS.

There are 5 main tests with subtests.

- TestPackage
-Package.PayloadFiles
-Package.PayloadFile
-Package.PayloadFile.DontExists
-Package.FootprintFile
- TestPackageManifest
-Package.Manifest.Stream
-Package.Manifest.Applications
-Package.Manifest.Properties
-Package.Manifest.Dependencies
-Package.Manifest.Capabilities
-Package.Manifest.Resources
-Package.Manifest.Tdf
-Package.Manifest.MsixDocument
-Package.Manifest.PackageId
- TestPackageBlockMap
-Package.BlockMap.Stream
-Package.BlockMap.ValidateFiles
- TestBundle
-Bundle.FootprintFile
-Bundle.PayloadPackages

- TestBundleManifest
-Bundle.Manifest.Stream
-Bundle.Manifest.PackageId
-Bundle.Manifest.PackageInfo

apitest.exe supports getting input from the command windows or taking a file with all the input. See test/api/input/apitest_test_1.cmakein for an example input file. This file is used in all automated tests. NOTE: don't use this file directly, use <bin folder>/test/api/input/apitest_test_1.txt
It is expected that the person who does the test knows what the input should be (eg. There is not: "Enter the package: " text output)

apitest.exe also supports redirecting the output to a file. That option is used for the AOSP and iOS BVTs.

Bugs found and fixed:
1. IAppxFile::GetName for CodeIntengrity footprint file returned AppxMetadata/CodeIntegrity.cat instead of AppxMetadata\CodeIntegrity.cat
1. IMsixElement::GetElements returned an empty IMsixElementEnumerator for AOSP.
1. S_OK and S_FALSE was taken as an int instead of signed long