Fix for new cmake tar command that won't take no files (#199)
Starting in cmake 3.15, the tar command will error if given no files to work on. This change will work across all version by cutting off that scenario and instead generating a 1 byte resource array. The code is then updated to recognize that scenario and short circuit to returning a file not found for any resources requested.