Add a test for packaging a file larger than 4GB (#203)
The primary change is the addition of a test that packs a file larger than 4GB. This is done by creating a stream that returns easily compressed data, resulting in an ~6MB package file. The test is tagged with a hidden [.slow] tag, as it takes ~5 minutes to churn through the data and then open the resulting package for verification.

The change also reorganizes the msix in directory into categories for current and future use. The categories (and directories) are:

public: headers that we "publish" for consumption by external callers
shared: headers used by the functional tests; no linking allowed
internal: headers used by future unit tests, were linking will happen
This allows the tests to reuse product functionality, rather than re-implementing things like ComPtr.