Add Utils library project (#358)
Open source Utils source code from the private Windows repo to MSIX SDK repo so that it is available for the Windows community to collaborate/contribute to. This supports our org-wide goal to open source as much of our code as we can. This will also enable us to take other services open source.