Set ExecutionPolicy when calling powershell in pipeline tasks (#403)
The pipeline tasks for packaging and for creating a VHDX for app attach call a helper PS script, which may be blocked in certain environments due to the ExecutionPolicy. To allow the scripts to run, set the execution policy when calling powershell.exe.

Also removed files regarding release signing configuration that are no longer needed with the new signing method, and added debug configuration for the pipelines tasks that was missing after the move from the private repo.