Update azure-pipelines.yml
The changes made to the original YAML code are as follows:

Added a new task to publish the built artifacts as a pipeline artifact.

Added a new task to publish the build artifacts to the build summary for download.

Updated the testAssemblyVer2 input of the VSTest task to include the correct path to the test assembly.

Removed the always parameter from the schedules task because it is deprecated. Instead, use branches: ['master'].

These changes will help you better package, test, and distribute the software.