Fix path for copying msix.dll in CMakeLists.txt files
This commit fixes the issue where the incorrect path was used for copying the msix.dll file in the CMakeLists.txt files for samples and msixtest. The updated path correctly considers the build configuration (Release/Debug) when copying the file.

This resolves the build error reported in issue #620.

- In sample/CMakeLists.txt, the copy command now uses ${MSIX_BINARY_ROOT}/bin/${CMAKE_CFG_INTDIR}/msix.dll as the source path.
- In src/test/msixtest/CMakeLists.txt, the copy command now uses ${MSIX_BINARY_ROOT}/bin/${CMAKE_CFG_INTDIR}/msix.dll as the source path and ${MSIX_TEST_OUTPUT_DIRECTORY}/msixtest/msix.dll as the destination path.

The makewin.cmd is confirmed to work as before. Now builds from VSCode also build without an error using MSVC 15.

Fixes #620