Fix inconsistent zip structure in .appx files created by makemsix.exe
This commit resolves issue #619 by ensuring that the central directory in .appx files created by makemsix.exe is sorted in the same order as MakeAppx.exe. The fix involves tracking the sequence of file names during the creation of the zip file and sorting the central directory entries accordingly before writing them to the stream.

- Add a vector to track the sequence of file names in ZipObjectWriter.
- Sort central directory entries based on the tracked file name sequence
  before writing them to the zip file.

Resolves: #619